refactor:concernsで可能の箇所を移動
