+++ app/models/gift_person.rb
@@ -1,4 +1,5 @@
 class GiftPerson < ApplicationRecord
+  include AvatarAttachable
   belongs_to :user
   belongs_to :relationship
   has_many :gift_records, foreign_key: "gift_people_id", dependent: :destroy
@@ -8,7 +9,6 @@ class GiftPerson < ApplicationRecord
   validates :name, presence: true, length: { minimum: 1 }
   validates :name, format: { with: /\A\S+.*\S*\z/, message: "空白のみは無効です" }
   validates :address, length: { maximum: 100 }, allow_blank: true
-  validate :avatar_validation
 
   # 指定された日付時点での年齢を計算
   def age_at(reference_date = Date.current)
@@ -45,39 +45,7 @@ class GiftPerson < ApplicationRecord
     end
   end
 
-  # ギフト相手のプロフィール画像関連メソッド
-  def avatar_url
-    return unless avatar.attached? && persisted?
-    begin
-      avatar
-    rescue ActiveRecord::RecordNotFound, NoMethodError => e
-      Rails.logger.warn "Gift person avatar URL generation failed: #{e.message}"
-      nil
-    end
-  end
-
-  def has_avatar?
-    avatar.attached?
-  end
-
-
   private
 
   # ギフト相手のプロフィール画像のバリデーション
-  def avatar_validation
-    return unless avatar.attached?
-
-    # ファイル形式チェック
-    unless avatar.content_type.in?(%w[image/jpeg image/jpg image/png image/webp])
-      errors.add(:avatar, "はJPEG、PNG、WEBP形式のファイルのみアップロードできます")
-    end
-
-    # ファイルサイズチェック（2MBまで）
-    if avatar.blob.byte_size > 2.megabytes
-      errors.add(:avatar, "のファイルサイズは2MB以下にしてください")
-    end
-  rescue StandardError => e
-    Rails.logger.error "Gift person avatar validation error: #{e.message}"
-    errors.add(:avatar, "の検証中にエラーが発生しました")
-  end
 end
