+++ app/views/static_pages/how_to_use.html.erb
@@ -4,7 +4,6 @@
   <!-- ページヘッダー -->
   <div class="w-full text-center py-16 bg-gradient-to-r from-blue-50 to-indigo-50 rounded-lg mb-12">
     <h1 class="text-4xl font-bold text-gray-800 mb-4">
-      <i class="fas fa-question-circle mr-3 text-blue-600"></i>
       使い方ガイド
     </h1>
     <p class="text-xl text-gray-600">
@@ -25,12 +24,6 @@
       <p class="text-gray-600 leading-relaxed mb-4">
         まずはアカウントを作成してログインしましょう。メールアドレスとパスワードだけで簡単に始められます。
       </p>
-      <div class="bg-blue-50 border-l-4 border-blue-400 p-4">
-        <p class="text-blue-800">
-          <i class="fas fa-info-circle mr-2"></i>
-          登録は無料で、個人情報は厳重に保護されます。
-        </p>
-      </div>
     </div>
 
     <!-- ステップ2 -->
